# feat: Performance Improvements and Accurate UI