--- a/src-tauri/Cargo.toml
+++ b/src-tauri/Cargo.toml
@@ -16,15 +16,17 @@ tauri-build = { version = "1.0.0-rc.5", features = [] }
 includedir_codegen = "0.6.0"
 
 [dependencies]
-serde_json = "1.0"
+serde_json = "1.0"  
 serde = { version = "1.0", features = ["derive"] }
 tauri = { version = "1.0.0-rc.5", features = ["api-all"] }
 runas = "0.2.1"
-phf = "0.8.0"
+phf = "0.11.1"
 includedir = "0.6.0"
-base64 = "0.12"
-rand_core = { version = "0.5", default-features = false, features = ["getrandom"] }
-x25519-dalek = "0.6"
+base64 = "0.13.0"
+rand_core = { version = "0.6.4", default-features = false, features = ["getrandom"] }
+x25519-dalek = "2.0.0-pre.1"
+boringtun = { version = "0.5.2", features = [] }
+zeroize = "1.5.7"
 
 [features]
 default = [ "custom-protocol" ]