--- a/src-tauri/src/tunnel/mod.rs
+++ b/src-tauri/src/tunnel/mod.rs
@@ -1,0 +1,3 @@
+mod peer;
+
+pub use peer::*;