--- a/src-tauri/src/tunnel/peer.rs
+++ b/src-tauri/src/tunnel/peer.rs
@@ -1,0 +1,11 @@
+
+fn peer() {
+    //      Make a way to connect peers using rust only
+    //      This will involve using boringtun
+
+    // let mut device_handle: DeviceHandle = match DeviceHandle::new(tun_name, config) {
+    //     Ok(d) => d,
+    //     Err(e) => {
+    //     }
+    // };
+}