--- a/src/components/reseda.tsx
+++ b/src/components/reseda.tsx
@@ -390,7 +390,7 @@ class WireGuard extends Component<{ file_path: string, user: any }> {
                         }
 
                         <div className="flex flex-row items-center justify-between">
-                            <p>Reseda FREE</p>
+                            <p>Reseda {this?.state?.tier}</p>
                             <p>{this?.user?.user?.email}</p>
 
                             <div onClick={() => { 
@@ -490,6 +490,7 @@ class WireGuard extends Component<{ file_path: string, user: any }> {
     }
 
     connect(location: Server, callback_time: Function) {
+        console.time("start->sendws");
         callback_time(new Date().getTime());
         console.log(this.config.keys);
         
@@ -509,6 +510,8 @@ class WireGuard extends Component<{ file_path: string, user: any }> {
                 query_type: "open"
             }));
 
+            console.timeEnd("start->sendws");
+
             this.setState({
                 ...this.state,
                 connection: {
@@ -520,7 +523,12 @@ class WireGuard extends Component<{ file_path: string, user: any }> {
                 }
             });
 
+            console.time("evt-listener->get-message");
+
             this.socket.addEventListener('message', async (connection) => {
+                console.timeEnd("evt-listener->get-message");
+                console.time("get-message->add-peer");
+
                 const connection_notes: Incoming = JSON.parse(connection.data);
 
                 if(connection_notes.type == "message" && typeof connection_notes.message == "object") {
@@ -538,7 +546,9 @@ class WireGuard extends Component<{ file_path: string, user: any }> {
                     });
     
                     this.socket.close();
-    
+                    
+                    console.timeEnd("get-message->add-peer");
+                    console.time("add-peer->finish");
                     await this.addPeer(location, message.server_public_key, message.subdomain, callback_time);
                     
                     this.setState({
@@ -553,6 +563,7 @@ class WireGuard extends Component<{ file_path: string, user: any }> {
                     });
     
                     callback_time(new Date().getTime());
+                    console.timeEnd("add-peer->finish");
                 }
             })
         }, location, true)
@@ -769,9 +780,11 @@ class WireGuard extends Component<{ file_path: string, user: any }> {
     }
 
     async up(cb: Function) {
+        console.time("up: checking wg");
         let k = await invoke('is_wireguard_up').then(e => {
             return `${e}`;
         });
+        console.timeEnd("up: checking wg");
 
         if(k.includes('RUNNING')) {
             await this.down(async () => {
@@ -787,9 +800,12 @@ class WireGuard extends Component<{ file_path: string, user: any }> {
     }
     
     async down(cb: Function) {
+        console.time("down: stopping wg");
         await invoke('stop_wireguard_tunnel', { path: this.config.wg.filePath }).then(e => {
             cb();
         })
+        console.timeEnd("down: stopping wg");
+
     }
 
     scrapeConfig() {