Now shows your current tier whilst also pulling your current usage maximums, allowing implementation of known disconnect reasoning, i.e. will not connect free or supporter tiers who have exceeded maximums and will be thrown by server.
Also improves connection times:
- 1000ms average reseda connection
- 2000ms average windows service startup

Resulting in an average **3s** connection time, on unix machines this can possibly be reduced. The base time requirement is the reseda connection, which depends heavily on the distance to the connecting server- i.e. at a 200ms ping-to-server, the connection takes 800ms to propagate and 200ms of computing time, thus 1000ms. The windows service startup is a standard time restraint that cannot be mitigated under the current implementation but should be amendable using a different method.